Create LogRule (#4559)
* Create LogRule

* Ktlint fix

* Android-required.
